requirements.txt
# requirements.txt
Flask==2.3.2
torch==2.0.1
torchvision==0.15.2
transformers==4.32.1
Pillow==10.0.0
requests==2.31.0
Werkzeug==2.3.6